feat: Add generic FindingGenerator and re-structure testing utilities
