+++ src/handler/index.ts
@@ -1,14 +1,7 @@
-
 import erc20Transfer from "./erc20.transfers";
 import ethTransfer from "./eth.transfers";
 import functionCalls from "./function.calls";
 import eventEmissions from "./events.checker";
 import blacklistedAddresses from "./blacklisted.addresses";
 
-export {
-    erc20Transfer,
-    ethTransfer,
-    functionCalls,
-    eventEmissions,
-    blacklistedAddresses,
-};
+export { erc20Transfer, ethTransfer, functionCalls, eventEmissions, blacklistedAddresses };
