+++ src/handler/types.ts
@@ -1,0 +1,3 @@
+import { Finding } from "forta-agent";
+
+export type FindingGenerator<T> = (data: T) => Finding;
